Move unban_peer test back
